requirements.txt
alpaca_trade_api==3.0.0
matplotlib==3.5.1
numpy
pandas==1.3.5
pandas_market_calendars==4.1.4
python_dateutil==2.8.1
Requests==2.31.0
seaborn==0.12.2
